requirements.txt
amqp==1.4.9
autopep8==1.2.4
Django==1.10
django-appconf==1.0.2
django-colorful==1.2
django-debug-toolbar==1.5
django-imagekit==3.3
django-phonenumber-field==1.1.0
django-twilio==0.8.0
djangorestframework==3.4.1
httplib2==0.9.2
numpy==1.11.1
pep8==1.7.0
phonenumberslite==7.5.1
pilkit==1.1.13
Pillow==3.3.0
pytz==2016.6.1
requests==2.10.0
six==1.10.0
sqlparse==0.2.0
twilio==5.4.0
virtualenv==15.0.2
